+++ tests/testthat/test-remotes.R
@@ -4,32 +4,29 @@ context("Remotes")
 test_that("we can parse a variety of remotes", {
 
   remote <- renv_remotes_parse("git@github.com:kevinushey/repo.git:subdir")
-
   expect_equal(remote$type,   "git")
   expect_equal(remote$host,   "github.com")
-  expect_equal(remote$user,   "kevinushey")
-  expect_equal(remote$repo,   "repo")
-  expect_equal(remote$ext,    "git")
+  expect_equal(remote$repo,   "kevinushey/repo.git")
   expect_equal(remote$subdir, "subdir")
 
   remote <- renv_remotes_parse("url::https://github.com/kevinushey/renv.git1/archive/refs/heads/main.zip")
-
   expect_equal(remote$type,     "url")
   expect_equal(remote$url,      "https://github.com/kevinushey/renv.git1/archive/refs/heads/main.zip")
   expect_equal(remote$protocol, "https")
 
   # https://github.com/rstudio/renv/issues/1004
   remote <- renv_remotes_parse("git@github.com:abc123/def456.git")
-
   expect_equal(remote$type,   "git")
   expect_equal(remote$host,   "github.com")
-  expect_equal(remote$user,   "abc123")
-  expect_equal(remote$repo,   "def456")
-  expect_equal(remote$ext,    "git")
+  expect_equal(remote$repo,   "abc123/def456.git")
 
   # https://github.com/rstudio/renv/issues/667
   remote <- renv_remotes_parse("package=git@github.com:abc123/def456.git")
   expect_equal(remote$package, "package")
+
+  # https://github.com/rstudio/renv/issues/667
+  remote <- renv_remotes_parse("git@github.com:abc/def/ghi")
+  expect_equal(remote$repo, "abc/def/ghi")
 
 })